refactor: Reorganize folders
