+++ src/main/java/linus/Main.java
@@ -8,6 +8,8 @@ import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
+import linus.gui.MainWindow;
+
 /**
  * A GUI for Linus using FXML.
  */
