+++ src/main/java/linus/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package linus;
+package linus.gui;
 
 import java.io.IOException;
 import java.util.Collections;
