+++ src/main/java/linus/gui/MainWindow.java
@@ -1,4 +1,4 @@
-package linus;
+package linus.gui;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -8,6 +8,8 @@ import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+
+import linus.Linus;
 
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
