+++ src/main/java/linus/storage/LocalDateTypeAdapter.java
@@ -1,4 +1,4 @@
-package linus.util;
+package linus.storage;
 
 import java.lang.reflect.Type;
 import java.time.LocalDate;
