+++ src/main/java/linus/storage/Storage.java
@@ -1,4 +1,4 @@
-package linus.util;
+package linus.storage;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -15,6 +15,8 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 
 import linus.exception.LinusException;
+import linus.storage.LocalDateTypeAdapter;
+import linus.storage.TaskDeserializer;
 import linus.task.Deadline;
 import linus.task.Event;
 import linus.task.Task;
